V8: use ES modules instead of global (#3361)
# Description of Changes

Tweaks V8 module support to use JS modules for `spacetimedb_sys` and the
user module rather than using scripts and the global object.

To this end, the code is also made more modular so that it e.g., cares
less avoid global vs. modules.
An `Object` with the functions is e.g., enough for the lowest level
`call_call_reducer`.

Some `.unwrap()`s are also removed.

Also, `run_timeout_and_cb_every` is disabled, as it currently leads to
UB due to bugs in the `v8` crate.

# API and ABI breaking changes

None

# Expected complexity level and risk

2

# Testing

Future work.