Compute read sets for MutTxId via InstanceEnv (#3520)
# Description of Changes

<!-- Please describe your change, mention any related tickets, and so on
here. -->

The `InstanceEnv` will now compute read sets when executing a view and
store them in the `MutTxId`.

These read sets track table scans as well as singular index key scans.
Index key ranges will be tracked in the future, but for now an index
range scan is treated as a full table scan.

These read sets are maintained as part of the `CommittedState`.

TODO: Check write sets against read sets. Re-evaluate views in the case
of overlap and update the read sets.

# API and ABI breaking changes

<!-- If this is an API or ABI breaking change, please apply the
corresponding GitHub label. -->

None

# Expected complexity level and risk

<!--
How complicated do you think these changes are? Grade on a scale from 1
to 5,
where 1 is a trivial change, and 5 is a deep-reaching and complex
change.

This complexity rating applies not only to the complexity apparent in
the diff,
but also to its interactions with existing and future code.

If you answered more than a 2, explain what is complex about the PR,
and what other components it interacts with in potentially concerning
ways. -->

1

# Testing

<!-- Describe any testing you've done, and any testing you'd like your
reviewers to do,
so that you're confident that all the changes work as expected! -->

- [ ] Unit tests