`Dockerfile` fails more cleanly if `.git` is a submodule (#3591)
# Description of Changes

Our `Dockerfile` runs `COPY . .`. If we're in a submodule rather than a
"full" git repo, that causes `.git` to be a in a weird state inside the
docker container, where it thinks it's in a submodule even though the
parent directory context has been lost.

This PR just adds a clear error message early in the process, because
this has bitten us a few times.

# API and ABI breaking changes

None

# Expected complexity level and risk

1

# Testing

- [x] We appropriately get the failure if we try to `docker build` this
from within a submodule.

Co-authored-by: Zeke Foppa <bfops@users.noreply.github.com>