Fix issue with row id for non primary key tables (#3603)
# Description of Changes

When computing rowId for non primary key tables the typescript sdk
incorrectly used the entire data buffer.
This resulted in rowIds being incorrect when multiple rows were present
within an update.

# Expected complexity level and risk

1

# Testing

Tested this in the context of developing the collaborator feature on the
spacetime.com website.
Without this change, updates to tables without primary keys were
sometimes not applied to the client cache.
With this update no further issues were observed.

Co-authored-by: Tyler Cloutier <cloutiertyler@users.noreply.github.com>