Typescript views (#3584)
# Description of Changes

Currently need changes from both #3548 and #3527, and the latter is
merged into master but the former isn't rebased on top of it. so we stay
silly for the moment

I'd like to pull out my first commit into its own PR, but that's not
really possible until #3548 rebases onto master.

# Expected complexity level and risk

2 - pretty straightforward translation of the wasm/Rust view
implementation to typescript.

# Testing

- [ ] <!-- maybe a test you want to do -->
- [ ] <!-- maybe a test you want a reviewer to do, so they can check it
off when they're satisfied. -->