Change `spacetime_worker_wasm_instance_errors_total` metric's labels (#3614)
# Description of Changes

There are two purposes for introducing the changes in this commit:

1. Decrease the cardinality of the metric, as the caller id and
connection id cwould result in way too combiations
2. Add a database_identity label, which is much more useful for the
metric

# API and ABI breaking changes

I'm honestly not sure if metrics labels change is an API breaking
change, I suppose not? For what it's worth, I don't think that anyone
was relying on the labels I've removed.

# Expected complexity level and risk

1

# Testing

- [x] I've tested the change locally

Co-authored-by: Tyler Cloutier <cloutiertyler@users.noreply.github.com>