Fixes docs links (#3803)
# Description of Changes

This removes all usages of `pathname:///` and replaces them with
`https://spacetimedb.com/`. `pathname:///` was resolving to `/docs` in
some cases and was not verified by our link checker. Note that this
means these links will NOT work in staging.

This is a temporary fix. We should find a way to be able to link out to
the rest of the site without hardcoding the hostname in the future.

# API and ABI breaking changes

None

# Expected complexity level and risk

1

# Testing

- I tested that the links work locally (but obviously they link to the
production site).

---------

Signed-off-by: John Detter <4099508+jdetter@users.noreply.github.com>
Co-authored-by: John Detter <4099508+jdetter@users.noreply.github.com>
Co-authored-by: joshua-spacetime <josh@clockworklabs.io>